+++ src/Aardvark.Data.Collada/RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+### 5.2.9
+- Updated to NET 8 and Aardvark.Base 5.3
+- 
+### 5.2.9-prerelease0001
+- Updated to NET 8 and Aardvark.Base 5.3 (prerelease)
+
 ### 5.2.8
 - Moved from Aardvark.Internal repository
 - Removed UnsafeCoerce() usage