+++ src/Aardvark.Data.Ifc/RELEASE_NOTES.md
@@ -8,6 +8,12 @@
 ### 1.1.0-prerelease0001
 - xBIM v6 dev packages
 
+### 1.0.4
+- Updated to NET 8 and Aardvark.Base 5.3
+
+### 1.0.4-prerelease0001
+- Updated to NET 8 and Aardvark.Base 5.3 (prerelease)
+
 ### 1.0.3
 - Updated Aardvark.Base packages