+++ src/Aardvark.Data.Photometry/RELEASE_NOTES.md
@@ -1,3 +1,12 @@
+### 5.3.10
+- Updated to NET 8 and Aardvark.Base 5.3
+- Fixed BuildEquidistantMatrix in case there is full 360° measurement data (#4855)
+- Added field documentation to LDTData
+- Review of non-equidistant fix (#4855)
+
+### 5.3.10-prerelease0001
+- Updated to NET 8 and Aardvark.Base 5.3 (prerelease)
+
 ### 5.3.9
 - Updated Aardvark.Base packages
 - Moved from Aardvark.Algodat