+++ src/Aardvark.PixImage.SystemDrawing/RELEASE_NOTES.md
@@ -1,2 +1,8 @@
+### 5.2.33
+- Updated to NET 8 and Aardvark.Base 5.3
+
+### 5.2.33-prerelease0001
+- Updated to NET 8 and Aardvark.Base 5.3 (prerelease)
+
 ### 5.2.32
 - Moved from Aardvark.Base